+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "reddsaver"
-version = "0.2.3"
+version = "0.2.4"
 authors = ["Manoj Karthick Selva Kumar <manojkarthick@ymail.com>"]
 description = "CLI tool to download saved images from Reddit"
 edition = "2018"
